* Reimplement Kst::DataRange so we can begin reading vectors in from disk.
git-svn-id: svn://anonsvn.kde.org/home/kde/branches/work/kst/portto4/kst@716917 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ src/libkstapp/vectordialog.h
@@ -33,11 +33,11 @@ class KST_EXPORT VectorTab : public DialogTab, Ui::VectorTab {
     void setMode(Mode mode) { _mode = mode; }
 
     //Slave vector mode methods...
-    int from() const;
-    void setFrom(int from);
+    qreal from() const;
+    void setFrom(qreal from);
 
-    int to() const;
-    void setTo(int to);
+    qreal to() const;
+    void setTo(qreal to);
 
     int numberOfSamples() const;
     void setNumberOfSamples(int numberOfSamples);
